Split getRawBrowserHistory out of blocking-history
